+++ src/client.rs
@@ -22,6 +22,7 @@ use valence_protocol::packets::s2c::play::{
     SetSubtitleText, SetTitleAnimationTimes, SetTitleText, SynchronizePlayerPosition,
     SystemChatMessage, UnloadChunk, UpdateAttributes, UpdateTime,
 };
+use valence_protocol::particle::{Particle, ParticleS2c};
 use valence_protocol::types::{
     AttributeProperty, DisplayedSkinParts, GameMode, GameStateChangeReason, SyncPlayerPosLookFlags,
 };
@@ -577,20 +578,20 @@ impl<C: Config> Client<C> {
 
     pub fn play_particle(
         &mut self,
-        particle_type: Particle,
-        position: Vec3<f64>,
-        offset: Vec3<f32>,
-        max_speed: f32,
-        particle_count: u32,
+        particle: &Particle,
         long_distance: bool,
+        position: impl Into<Vec3<f64>>,
+        offset: impl Into<Vec3<f32>>,
+        max_speed: f32,
+        count: i32,
     ) {
         self.queue_packet(&ParticleS2c {
-            particle_type,
+            particle: particle.clone(),
             long_distance,
-            position,
-            offset,
+            position: position.into().into_array(),
+            offset: offset.into().into_array(),
             max_speed,
-            particle_count,
+            count,
         })
     }
